refactor: move gaspriceconfig to saetypes
