+++ gastime/gastime.go
@@ -15,6 +15,7 @@ import (
 	"github.com/ava-labs/strevm/hook"
 	"github.com/ava-labs/strevm/intmath"
 	"github.com/ava-labs/strevm/proxytime"
+	saetypes "github.com/ava-labs/strevm/types"
 )
 
 //go:generate go run github.com/StephenButtolph/canoto/canoto $GOFILE
@@ -44,7 +45,7 @@ type Time struct {
 // New returns a new [Time], derived from a [time.Time]. The consumption of
 // `target` * [TargetToRate] units of [gas.Gas] is equivalent to a tick of 1
 // second.
-func New(at time.Time, target, startingExcess gas.Gas, gasPriceConfig hook.GasPriceConfig) (*Time, error) {
+func New(at time.Time, target, startingExcess gas.Gas, gasPriceConfig saetypes.GasPriceConfig) (*Time, error) {
 	cfg, err := newConfig(gasPriceConfig)
 	if err != nil {
 		return nil, err
@@ -88,9 +89,9 @@ const DefaultTargetToExcessScaling = 87
 // parameter in ACP-176's price calculation.
 const DefaultMinPrice gas.Price = 1
 
-// DefaultGasPriceConfig returns the default [hook.GasPriceConfig] values.
-func DefaultGasPriceConfig() hook.GasPriceConfig {
-	return hook.GasPriceConfig{
+// DefaultGasPriceConfig returns the default [saetypes.GasPriceConfig] values.
+func DefaultGasPriceConfig() saetypes.GasPriceConfig {
+	return saetypes.GasPriceConfig{
 		TargetToExcessScaling: DefaultTargetToExcessScaling,
 		MinPrice:              DefaultMinPrice,
 		StaticPricing:         false,
@@ -138,8 +139,8 @@ func (tm *Time) Excess() gas.Gas {
 }
 
 // Price returns the price of a unit of gas, i.e. the "base fee", determined by
-// [gas.CalculatePrice]. However, when [hook.GasPriceConfig.StaticPricing] is
-// true, Price always returns [hook.GasPriceConfig.MinPrice].
+// [gas.CalculatePrice]. However, when [saetypes.GasPriceConfig.StaticPricing] is
+// true, Price always returns [saetypes.GasPriceConfig.MinPrice].
 func (tm *Time) Price() gas.Price {
 	if tm.config.staticPricing {
 		return tm.config.minPrice
