+++ hook/hookstest/stub.go
@@ -33,7 +33,7 @@ type Stub struct {
 	Ops                     []Op
 	ExecutionResultsDBFn    func(string) (saetypes.ExecutionResults, error)
 	CanExecuteTransactionFn func(common.Address, *common.Address, libevm.StateReader) error
-	GasPriceConfig          hook.GasPriceConfig
+	GasPriceConfig          saetypes.GasPriceConfig
 }
 
 var _ hook.PointsG[Op] = (*Stub)(nil)
@@ -42,7 +42,7 @@ var _ hook.PointsG[Op] = (*Stub)(nil)
 type HookOption = options.Option[Stub]
 
 // WithGasPriceConfig overrides the default gas config.
-func WithGasPriceConfig(cfg hook.GasPriceConfig) HookOption {
+func WithGasPriceConfig(cfg saetypes.GasPriceConfig) HookOption {
 	return options.Func[Stub](func(s *Stub) {
 		s.GasPriceConfig = cfg
 	})
@@ -74,7 +74,7 @@ func WithExecutionResultsDBFn(fn func(string) (saetypes.ExecutionResults, error)
 func NewStub(target gas.Gas, opts ...HookOption) *Stub {
 	// defaultGasPriceConfig is the same as [gastime.DefaultGasPriceConfig]. It is defined
 	// here to avoid a circular dependency between [gastime] and [hookstest].
-	defaultGasPriceConfig := hook.GasPriceConfig{
+	defaultGasPriceConfig := saetypes.GasPriceConfig{
 		TargetToExcessScaling: 87,
 		MinPrice:              1,
 		StaticPricing:         false,
@@ -178,7 +178,7 @@ func (s *Stub) BlockRebuilderFrom(b *types.Block) (hook.BlockBuilder[Op], error)
 }
 
 // GasConfigAfter ignores its argument and always returns [Stub.Target] and [Stub.GasPriceConfig].
-func (s *Stub) GasConfigAfter(*types.Header) (gas.Gas, hook.GasPriceConfig) {
+func (s *Stub) GasConfigAfter(*types.Header) (gas.Gas, saetypes.GasPriceConfig) {
 	return s.Target, s.GasPriceConfig
 }
